+++ common/src/main/java/io/github/sds100/keymapper/common/utils/MapUtils.kt
@@ -3,5 +3,4 @@ package io.github.sds100.keymapper.common.utils
 /**
  * Not for high speed stuff.
  */
-fun <K, V> Map<K, V>.getKey(value: V) =
-    entries.firstOrNull { it.value == value }?.key
+fun <K, V> Map<K, V>.getKey(value: V) = entries.firstOrNull { it.value == value }?.key