+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - #727 Actions to send SMS messages: "Send SMS" and "Compose SMS"
 - #1819 Explain how to enable the accessibility service restricted setting
 - #661 Action to execute shell commands.
+- #991 Consolidated volume and stream actions.
+- #1066 Action to mute/unmute microphone.
 
 ## Removed
 
@@ -42,7 +44,8 @@
 - #1818 the Key Mapper GUI Keyboard is no longer mentioned in the app. It still works but PRO mode
   and the auto switching feature are the preferred way to work around the limitations of the Key
   Mapper keyboard.
-- Allow selecting notification and alarm sound and not just ringtones for Sound action
+- Allow selecting notification and alarm sound and not just ringtones for Sound action.
+- #1064 wait for switch keyboard action to complete before doing next action.
 
 ## [3.2.1](https://github.com/sds100/KeyMapper/releases/tag/v3.2.1)